Code refactoring : Moved constants to constant file.
