+++ src/screens/work-interest/index.js
@@ -2,124 +2,8 @@ import { useState } from 'react';
 import { ProgressBar, Radio, Button, MultiSelect } from '../../components';
 import User from '../../models/user';
 import { useHistory } from "react-router-dom";
+import { industryPreferredForWork } from '../../constants';
 
-const industryPreferredForWork = [
-    {
-        label: 'Construction',
-        value: 'Construction',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Hospitality',
-        value: 'Hospitality',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Real estate',
-        value: 'Real estate',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Tourism',
-        value: 'Tourism',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Manufacturing',
-        value: 'Manufacturing',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Farming',
-        value: 'Farming',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Fishing',
-        value: 'Fishing',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Education',
-        value: 'Education',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Public services',
-        value: 'Public services',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Administration',
-        value: 'Administration',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Media',
-        value: 'Media',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Technology',
-        value: 'Technology',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Financial services',
-        value: 'Financial services',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Legal',
-        value: 'Legal',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Health care',
-        value: 'Health care',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Government',
-        value: 'Government',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Medical',
-        value: 'Medical',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Security',
-        value: 'Security',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Logistic',
-        value: 'Logistic',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Robotics',
-        value: 'Robotics',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Automotive',
-        value: 'Automotive',
-        location: '/images/family.png'
-    },
-    {
-        label: 'Transportation',
-        value: 'Transportation',
-        location: '/images/calender.png'
-    },
-    {
-        label: 'Customer support',
-        value: 'Customer support',
-        location: '/images/calender.png'
-    }
-]
 
 const WorkInterest = () => {
